refactor: PlanType에서 Plan으로 변경 및 테스트 코드 수정
1. PlanType을 Plan 객체로 대체하여 모든 관련 테스트 코드를 수정한다.
2. Users 및 QueryLog 관련 테스트에서 Plan 객체 사용 방식으로 변경한다.
3. 테스트 데이터 생성 시 Plan의 정적 팩토리 메서드를 활용하도록 적용한다.

Resolves #14
