+++ src/test/java/com/posicube/assignment/domain/querylog/QueryLogTest.java
@@ -1,12 +1,12 @@
 package com.posicube.assignment.domain.querylog;
 
 import com.posicube.assignment.common.exception.BaseException;
-import com.posicube.assignment.plan.domain.entity.PlanType;
+import com.posicube.assignment.plan.domain.entity.Plan;
+import com.posicube.assignment.plan.domain.vo.PlanType;
 import com.posicube.assignment.querylog.domain.ModelType;
 import com.posicube.assignment.querylog.domain.QueryLog;
 import com.posicube.assignment.querylog.exception.QueryLogExceptionStatus;
 import com.posicube.assignment.users.domain.entity.Users;
-import com.posicube.assignment.users.presentation.dtos.UserCreateRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -22,7 +22,9 @@ public class QueryLogTest {
 
     @BeforeEach
     void setUp() {
-        testUser = Users.create("evanbackeng@gmail.com", "pass1234", "evan", PlanType.PRO);
+        testUser = Users.create(
+                "evanbackeng@gmail.com", "pass1234", "evan", Plan.create(PlanType.PRO)
+        );
     }
 
     @Test
