+++ src/test/java/com/posicube/assignment/domain/users/UsersTest.java
@@ -1,7 +1,8 @@
 package com.posicube.assignment.domain.users;
 
 import com.posicube.assignment.common.exception.BaseException;
-import com.posicube.assignment.plan.domain.entity.PlanType;
+import com.posicube.assignment.plan.domain.entity.Plan;
+import com.posicube.assignment.plan.domain.vo.PlanType;
 import com.posicube.assignment.users.domain.entity.Users;
 import com.posicube.assignment.users.exception.UserExceptionStatus;
 import org.junit.jupiter.api.BeforeEach;
@@ -19,7 +20,7 @@ public class UsersTest {
         String account = "evanbackeng@gmail.com";
         String password = "pass1234";
         String name = "evan";
-        PlanType plan = PlanType.LITE;
+        Plan plan = Plan.create(PlanType.LITE);
         user = Users.create(account, password, name, plan);
     }
 
@@ -37,15 +38,18 @@ public class UsersTest {
     @DisplayName("생성 실패: User account, password, name이 null이면 예외 발생")
     public void create_user_account_password_name_cannot_null_fail() throws Exception {
         //when&then
-        assertThatThrownBy(() -> Users.create(null, "pass1234", "evan", PlanType.LITE))
+        assertThatThrownBy(() -> Users.create(
+                null, "pass1234", "evan", Plan.create(PlanType.LITE)))
                 .isInstanceOf(BaseException.class)
                 .hasMessage(UserExceptionStatus.ACCOUNT_CANNOT_BE_NULL.getMessage());
 
-        assertThatThrownBy(() -> Users.create("evanbackeng@gmail.com", null, "evan", PlanType.LITE))
+        assertThatThrownBy(() -> Users.create(
+                "evanbackeng@gmail.com", null, "evan", Plan.create(PlanType.LITE)))
                 .isInstanceOf(BaseException.class)
                 .hasMessage(UserExceptionStatus.PASSWORD_CANNOT_BE_NULL.getMessage());
 
-        assertThatThrownBy(() -> Users.create("evanbackeng@gmail.com", "pass1234", null, PlanType.LITE))
+        assertThatThrownBy(() -> Users.create(
+                "evanbackeng@gmail.com", "pass1234", null, Plan.create(PlanType.LITE)))
                 .isInstanceOf(BaseException.class)
                 .hasMessage(UserExceptionStatus.NAME_CANNOT_BE_NULL.getMessage());
     }
@@ -54,15 +58,18 @@ public class UsersTest {
     @DisplayName("생성 실패: User account, password, name이 공백이 포함되면 예외 발생")
     public void create_user_account_password_name_cannot_white_space_fail() throws Exception {
         //when&then
-        assertThatThrownBy(() -> Users.create("ev an  backeng@  gmail.    com", "pass1234", "evan", PlanType.LITE))
+        assertThatThrownBy(() -> Users.create(
+                "ev an  backeng@  gmail.    com", "pass1234", "evan", Plan.create(PlanType.LITE)))
                 .isInstanceOf(BaseException.class)
                 .hasMessage(UserExceptionStatus.ACCOUNT_CANNOT_CONTAIN_WHITESPACE.getMessage());
 
-        assertThatThrownBy(() -> Users.create("evanbackeng@gmail.com", "p as  s12   34", "evan", PlanType.LITE))
+        assertThatThrownBy(() -> Users.create(
+                "evanbackeng@gmail.com", "p as  s12   34", "evan", Plan.create(PlanType.LITE)))
                 .isInstanceOf(BaseException.class)
                 .hasMessage(UserExceptionStatus.PASSWORD_CANNOT_CONTAIN_WHITESPACE.getMessage());
 
-        assertThatThrownBy(() -> Users.create("evanbackeng@gmail.com", "pass1234", "e v  a   n", PlanType.LITE))
+        assertThatThrownBy(() -> Users.create(
+                "evanbackeng@gmail.com", "pass1234", "e v  a   n", Plan.create(PlanType.LITE)))
                 .isInstanceOf(BaseException.class)
                 .hasMessage(UserExceptionStatus.NAME_CANNOT_CONTAIN_WHITESPACE.getMessage());
     }
